- Finished support for partial loading of neighborhoods (finally!); - Fixed a bug with byte graph decoder.
git-svn-id: svn://localhost/tmp/repo1/trunk@339 09e1528e-719b-4268-a937-68dd4ef2d4b4

+++ Shared/main/src/it/unitn/disi/graph/codecs/ByteGraphDecoder.java
@@ -34,7 +34,7 @@ public class ByteGraphDecoder extends AbstractEdgeListDecoder {
 			}
 			return -1;
 		} else if (read != fBuf.length) {
-			unexpectedEOF("file size is not multiple of " + Long.SIZE);
+			unexpectedEOF("file size is not multiple of " + Integer.SIZE);
 		}
 
 		return CodecUtils.decodeInt(fBuf);
